requirements.txt
streamlit>=1.31.0
pandas>=2.1.0
numpy>=1.24.0
plotly>=5.18.0
pillow>=10.0.0
openai>=1.10.0
anthropic>=0.18.0
python-dotenv>=1.0.0
colorthief>=0.2.1
scikit-learn>=1.3.0
scipy>=1.11.0
requests>=2.31.0
